Fix bug where auth screen appears black on some systems
I'm not sure why, but apparently some systems or versions of xsecurelock pass the
root window as `XSCREENSAVER_WINDOW`, and others will pass the child window. Based
on the implementation of `auth_x11` in the xsecurelock source, it looks like we should
always assume this is a child window, so we should query the parent window when
creating our child window.